app/requirements.txt
Flask==3.0.0
flask[async]
Flask-PyMongo==2.3.0
Flask-Cors==4.0.0
pyJWT
python-dotenv==1.0.1
pymongo==4.6.3